# perf: reduce bundle size