--- a/packages/yarnpkg-builder/sources/commands/build/bundle.ts
+++ b/packages/yarnpkg-builder/sources/commands/build/bundle.ts
@@ -120,7 +120,21 @@ export default class BuildBundleCommand extends Command {
           },
           entryPoints: [path.join(basedir, `sources/cli.ts`)],
           bundle: true,
-          define: {YARN_VERSION: JSON.stringify(version)},
+          define: {
+            YARN_VERSION: JSON.stringify(version),
+            ...(this.noMinify ? {} : {
+              // For React
+              'process.env.NODE_ENV': JSON.stringify(`production`),
+              // For ink
+              'process.env.DEV': JSON.stringify(`false`),
+              // For ourselves
+              'process.env.TEST_ENV': `false`,
+              // mkdirp
+              'process.env.__TESTING_MKDIRP_PLATFORM__': `false`,
+              'process.env.__TESTING_MKDIRP_NODE_VERSION__': `false`,
+              'process.env.__FAKE_PLATFORM__': `false`,
+            }),
+          },
           outfile: output,
           logLevel: `silent`,
           format: `iife`,