**What's the problem this PR addresses?**

The bundle size increased from `1 973 101` bytes to `3 282 027` bytes in https://github.com/yarnpkg/berry/pull/4253.

**How did you ~~fix~~ improve it?**

- Patch `ink` to only import the `lodash` method it needs.
- Define some `process.env.*` values so ESBuild can do some more tree shaking.

**Results**

|     | Size in bytes |
| --- | ---- |
|Master | 3 288 047 |
|Patch `ink` | 3 209 390 |
|Define `process.env.NODE_ENV` | 3 004 912 |
|Define `process.env.DEV` | 2 794 423 |
|Define `process.env.TEST_ENV` | 2 794 247 |
|Define misc. `mkdirp` test variables | 2 794 130 |

```sh
YARN_IGNORE_PATH=1 hyperfine -w 5 "node ./master.cjs --version" "node ./after.cjs --version"
Benchmark 1: node ./master.cjs --version
  Time (mean ± σ):     218.1 ms ±   1.3 ms    [User: 227.8 ms, System: 22.2 ms]
  Range (min … max):   215.2 ms … 219.6 ms    13 runs

Benchmark 2: node ./after.cjs --version
  Time (mean ± σ):     207.6 ms ±   1.6 ms    [User: 227.4 ms, System: 12.2 ms]
  Range (min … max):   205.0 ms … 210.0 ms    14 runs

Summary
  'node ./after.cjs --version' ran
    1.05 ± 0.01 times faster than 'node ./master.cjs --version'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.