@trivikr 2022-04-04
> * Patch `ink` to only import the `lodash` method it needs

Is there an issue PR upstream for merging this fix?

I see that the source code is already doing named imports, probably some issue with transpilation
https://github.com/vadimdemedes/ink/blob/4fa00de3a0e8b96562c85d3b65673be0dc15ffc2/src/ink.tsx#L2