@merceyz 2022-04-06
> Is there an issue or PR upstream for merging this fix?

There is now https://github.com/vadimdemedes/ink/pull/511

> I see that the source code is already doing named imports, probably some issue with transpilation

That doesn't matter in this case as `lodash` isn't tree shakeable.